extracted ranged attack handlers
