+++ components/consumable.py
@@ -6,7 +6,9 @@ from actions.swap_action import SwapAction
 from components.base_component import BaseComponent
 from components.inventory import Inventory
 from exceptions import Impossible
-from handlers.input_handlers import ActionOrHandler, AreaRangedAttackHandler, SelectAdjacentHandler, SingleRangedAttackHandler
+from handlers.area_ranged_attack_handler import AreaRangedAttackHandler
+from handlers.single_ranged_attack_handler import SingleRangedAttackHandler
+from handlers.input_handlers import ActionOrHandler, SelectAdjacentHandler
 import actions.item_action
 import color
 import components.ai.confused_ai
