+++ game_map.py
@@ -129,9 +129,13 @@ class GameWorld:
     def generate_floor(self) -> None:
         import procgen.boss
         import procgen.rooms
+        import procgen.devroom
 
         self.current_floor += 1
-        if self.current_floor in [2, 4, 6]:
+        if self.current_floor == 1:
+            self.engine.game_map = procgen.devroom.generate(self.engine)
+            return
+        elif self.current_floor in [2, 4, 6]:
             self.engine.game_map = procgen.boss.generate(self.engine)
             return
         elif self.current_floor == 7:
