+++ handlers/single_ranged_attack_handler.py
@@ -1,0 +1,19 @@
+from actions.action import Action
+from engine import Engine
+from handlers.input_handlers import SelectIndexHandler
+from typing import Optional
+
+
+from typing import Callable, Tuple
+
+
+class SingleRangedAttackHandler(SelectIndexHandler):
+    """Handles targeting a single enemy. Only the enemy selected will be affected."""
+
+    def __init__(self, engine: Engine, callback: Callable[[Tuple[int, int]], Optional[Action]]):
+        super().__init__(engine)
+
+        self.callback = callback
+
+    def on_index_selected(self, x: int, y: int) -> Optional[Action]:
+        return self.callback((x, y))
